Update project docs with name "karmanaut"
This is a forked version of my project "chameleon", which samples
Stack Overflow reputation over time. Karmanaut, on the other hand,
is designed to sample Reddit karma over time. The projects are
very similar, but the specifics are different.

Hopefully I can eventually factor out a common base for both projects
to use; for now, I'll just use Chameleon as the basis for Karmanaut.