Update Karmanaut's user-agent string
Karmanaut was not using a proper user-agent string, and so the Reddit
API was returning HTTP 429 for all requests.

I should probably fix this so that my client (i.e., the one I'm running
for my own collection) is using a unique string, and the open-source
version uses another, because it's possible other users are also using
Karmanaut...but this works for now.