Properly embed and retrieve version number
The way I was doing it before required project.clj to be in the current
working directory when running the .jar file (i.e., it did not pull out
the project.clj file from the .jar). Apparently the best way, as noted
in [1], is to embed the version number in the JAR manifest file and
retrieve it that way.

I modified utils/version to do it both ways, so that it will work
without building a .jar file _and_ when running the crawler from a .jar
file.

  [1] http://stackoverflow.com/a/12599890/28804